complite expansion, add separate c-files for expand, redirect, utils
