+++ utils_functions.c
@@ -1,0 +1,59 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   utils_functions.c                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: epolkhov <epolkhov@student.42.fr>          #+#  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024-07-02 16:22:56 by epolkhov          #+#    #+#             */
+/*   Updated: 2024-07-02 16:22:56 by epolkhov         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+#include "builtins.h"
+
+char *ft_strncpy(char *dest, const char *src, size_t n)
+{
+	size_t  i;
+
+	i = 0;
+	while (i < n && src[i] != '\0')
+	{
+		dest[i] = src[i];
+		i++;
+	}
+	while (i < n)
+	{
+		dest[i] = '\0';
+		i++;
+	}
+
+	return dest;
+}
+
+char *ft_strndup(const char *str, size_t n)
+{
+	size_t len;
+	size_t i;
+	char *dup;
+
+	len =  0;
+	i = 0;
+	while (str[i] && len < n)
+	{
+		len++;
+		i++;
+	}
+	dup = (char *)malloc(len + 1);
+	if (!dup)
+		return (NULL);
+	i = 0;
+	while (i < len)
+	{
+		dup[i] = str[i];
+		i++;
+	}
+	dup[i] = '\0';
+	return (dup);
+}
